Used landfill point data to create 1km buffer around each
Because methane would be emitted as a plume that will drift away from the landfill, an area around each landfill is needed to capture any emissions. QGIS was used to created buffered landfills and then this was saved as a GeoJSON as I think this is likely the most acceptable format for the geometries used in the Copernicus NDVI tool I'm repurposing. Still waiting for a response from them regarding the exact method.