Buffered geometry not currently working as intended
When the point data is buffered, only 3 of the 23 geometries seemingly have data. I've checked the long/lat of about 10 points of the buffered gometry and they all seem fine. Point data when loaded accurately reflets 23 locations.

I'm waiting for a response from Copernicus as to what the problem could be. For now, I'm loading point data of the landfills and moving on with coding the rest.