Added module that downloads NetCDF
This bit of code can download the .nc (NetCDF) file for a specified date, or a series of dates. The plan is to display this on a map somehow so that the end user can investigate a specific methane spike from the time series and then assess if it likely comes from their landfill.