The methane map using Sentinel 2 data now works.
I found this code that extracted the methane data from sentinel 2 imagery using a method not dissimilar to my own.

https://github.com/aws/amazon-sagemaker-examples/blob/main/sagemaker-geospatial/methane-emission-monitoring/monitor_methane_ch4_emission_point_sources.ipynb

I have taken the code within and adapted it to replace the one I originally created for the single band multi-pass method.

A site with a known methane emission was taken near Madrid to serve as the test
https://www.theguardian.com/environment/2024/feb/12/revealed-the-1200-big-methane-leaks-from-waste-dumps-trashing-the-planet

And the methane clouds successfully displayed along with what seems to be the apparent source.